requirements.txt
PySide6
python-dotenv
matplotlib
numpy
pyserial